isolates and moves xslt package.
