+++ src/main/java/nl/kb/dare/oai/IndexMetadataTask.java
@@ -7,7 +7,7 @@ import nl.kb.http.HttpFetcher;
 import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.model.repository.Repository;
 import nl.kb.dare.model.repository.RepositoryDao;
-import nl.kb.dare.xslt.PipedXsltTransformer;
+import nl.kb.xslt.PipedXsltTransformer;
 
 import java.io.PrintWriter;
 import java.util.List;
