+++ src/main/java/nl/kb/dare/oai/ListRecords.java
@@ -5,7 +5,7 @@ import nl.kb.http.HttpFetcher;
 import nl.kb.http.responsehandlers.ResponseHandlerFactory;
 import nl.kb.dare.model.repository.Repository;
 import nl.kb.dare.model.statuscodes.OaiStatus;
-import nl.kb.dare.xslt.PipedXsltTransformer;
+import nl.kb.xslt.PipedXsltTransformer;
 import org.xml.sax.InputSource;
 
 import javax.xml.transform.stream.StreamResult;
